+++ src/lib/sveltekit_shim_env.test.ts
@@ -1,5 +1,4 @@
-import {test} from 'uvu';
-import * as assert from 'uvu/assert';
+import {describe, test, expect} from 'vitest';
 import {resolve} from 'node:path';
 import {init_test_env} from './test_helpers.ts';
 
@@ -9,15 +8,15 @@ const VALUE = 'SOME_PUBLIC_ENV_VAR';
 
 // dynamic import paths are needed to avoid building .d.ts and .d.ts.map files, could be fixed in the build process
 
-test('shims static SvelteKit $env imports', async () => {
-	const mod = await import(resolve('src/fixtures/test_sveltekit_env.ts'));
-	assert.is(mod.exported_env_static_public, VALUE);
-});
+describe('sveltekit shim env', () => {
+	test('shims static SvelteKit $env imports', async () => {
+		const mod = await import(resolve('src/fixtures/test_sveltekit_env.ts'));
+		expect(mod.exported_env_static_public).toBe(VALUE);
+	});
 
-test('shims dynamic SvelteKit $env imports', async () => {
-	const mod = await import('$env/static/public');
-	// @ts-ignore
-	assert.is(mod.PUBLIC_SOME_PUBLIC_ENV_VAR, VALUE);
+	test('shims dynamic SvelteKit $env imports', async () => {
+		const mod = await import('$env/static/public');
+		// @ts-ignore
+		expect(mod.PUBLIC_SOME_PUBLIC_ENV_VAR).toBe(VALUE);
+	});
 });
-
-test.run();